Finally I decided to add High Performance SSH. http://www.psc.edu/networking/projects/hpn-ssh/
I tried to keep the openssh hash not changing, unless the user sets hpn in getConfig
style. I think that does not look as good as a patch changing the hash, but it may
annoy less. Let me know if it is not ok.

I don't think hpn should be the default, because it may have some insecurity implications
I don't know of. But I used to enable it in all my machines, and I hope to do so unless
advised otherwise.

git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@18073 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb