pkgs/development/libraries/sqlite/default.nix: updated to version 3.6.19
The new version requires Tcl at build time in order to construct sqlite3.h,
even when the --without-tcl option is passed to configure. That feels odd
because the sqlite web site does advertise "no dependencies", but then it's
probably not a big deal either. The build of the Tcl plugin for sqlite is still
disabled, so there is no run-time dependency.

git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@18091 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb