Add coq-8.3rc1
Note: In this version we introduce a new schema for the name of the coq
derivations where the coq version is included in the name (i.e.,
"coq8.3-8.3pre1" instead of "coq-8.3pre1").  The reason for this is that often
coq releases introduce several incompatibilities.  Thus I argue that, in
general, users do not want nix-env to upgrade automatically form one release to
another.  Also version string "8.3pre1" is used instead of "8.3-rc1" to trigger
the nix mechanism for versions comparison.



git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@23803 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb