Dwarf Fortress
The wrapper script works by creating an image of the distribution in the users home directory
(defaulting to $HOME/.config/df_linux).  This is done because the program expects to write
state into it's own distribution.

Better ideas are welcome!

What further is needed is to allow users to configure the init.  One way is to use nix-config
to apply patches to the init.txt file, though this would break sharing of the package between
different users.



git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@24538 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb