pkgs/misc/emulators/wine: enable Hydra builds on x86_64-linux
Wine is supposed to compile on both 32 and 64-bit Linux. When I try to build
it in a 64 bit environment, however, the build fails while compiling libgmp,
saying that the C compiler in bootstrap-tools doesn't support C++. I'm not
sure why that happens. Hopefully, the builds in Hydra give me some clue
about this phenomenon.

git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@25693 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb