Fixed rtl8192cfw firmware package.
Also, added a package that fetches all firmware from dwmw2's repo on kernel.org.
Specific firmware can then simply copy the driver out of this repo. Perhaps it
would actually be useful to write some abstraction for that.

Anyway, I'm not sure if fetching the git repo is a good idea to start with. The
repo does not have tags, so I'm forced to either take the HEAD or a particular
commit (chose the last option). However, this is likely to need ongoing
maintenance.


git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@31492 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb