intel X driver: upgraded to fix display glitches on newer chips
at least 2 nixos users were affected by a graphics bug that got fixed
in 2.15.
Officially 2.15 was meant for a newer xorg server, libdrm, mesa and all, but
I've been running with this for over a month with no issues (used
compositing desktop, opengl and hd video).

I know 2.17 is in the repo, but it doesn't builde (missing deps I
think).

git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@31661 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb