Making libreoffice have a configured fontconfig. I hope this will make some cppunit tests work, and thus we would not need to disable them.
I update the patch that disables tests with one more test disabled, but I don't
use it because I hope the build with fonts.conf will work fine. If that is the case,
I'll remove the patch file later. If the fonts.conf trick does not work, I may want
to extend the patch to disable all failing cases.



git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@33062 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb